+++ utils/constants.js
@@ -3,6 +3,7 @@ module.exports = Object.freeze({
     DB_CDS : 'cap-valid-data-model.cds',
     DB_DATA_FOLDER : 'data',
     SRV_FOLDER : 'srv',
+    TEST_FOLDER : 'test',
     SRV_EXT_FOLDER : 'external',
     BOOKS_CSV : 'my.validexample-Books.csv',
     AUTHORS_CSV : 'my.validexample-Authors.csv',
@@ -17,5 +18,11 @@ module.exports = Object.freeze({
     XS_SECURITY : 'xs-security.json',
     XS_SECURITY_CC : 'xs-security-cc.json',
     MTA : 'mta.yaml',
-    SAP_HTTP_CLIENT : '@sap-cloud-sdk/http-client'
+    SAP_HTTP_CLIENT : '@sap-cloud-sdk/http-client',
+    TEST_DEPENDENCIES : "npm i -D mocha chai@4 chai-as-promised chai-subset",
+    UI5_MODULE : "[ui5]",
+    EXT_MODULE : "[ext]",
+    MTA_MODULE : "[mta]",
+    TEST_MODULE : "[test]",
+    PROFILING_MODULE : "[profiling]"
 });