+++ utils/external.js
@@ -14,7 +14,7 @@ const constants = require('./constants');
 let createExtFolder = function(target){
     return new Promise(async (resolve, reject) => {
         try {
-            logger.info(`Copying ${constants.SRV_FOLDER}/${constants.SRV_EXT_FOLDER} and its content...`);
+            logger.info(`${constants.EXT_MODULE} - Copying ${constants.SRV_FOLDER}/${constants.SRV_EXT_FOLDER} and its content...`);
 
             await fsExtra.ensureDir(path.join(target,constants.SRV_FOLDER,constants.SRV_EXT_FOLDER));
 
@@ -23,10 +23,10 @@ let createExtFolder = function(target){
                             path.join(target,constants.SRV_FOLDER,constants.SRV_EXT_FOLDER)
                         );
         
-            logger.info(`Folder ${constants.SRV_FOLDER}/${constants.SRV_EXT_FOLDER} and its content copied`);
+            logger.info(`${constants.EXT_MODULE} - Folder ${constants.SRV_FOLDER}/${constants.SRV_EXT_FOLDER} and its content copied`);
             resolve();
           } catch (err) {
-            logger.error('Error during folder copy:', err);
+            logger.error('${constants.EXT_MODULE} - Error during folder copy:', err);
             reject(err);
           }
     })
@@ -43,15 +43,15 @@ let modifyPackage = function (target) {
             let sourcePackage = editJsonFile(path.join(__dirname, '..', constants.PACKAGE_CAP));
             let nwSource = await sourcePackage.get('cds.requires.nw');
             targetPackage.set('cds.requires.nw', nwSource);
-            logger.info(`External source nortwhind added in /${constants.PACKAGE}`);
+            logger.info(`${constants.EXT_MODULE} - External source nortwhind added in /${constants.PACKAGE}`);
             let sapHttpClient = await targetPackage.get(`dependencies.${constants.SAP_HTTP_CLIENT}`);
             if(!sapHttpClient) {
-                logger.warn(`${constants.SAP_HTTP_CLIENT} dependency is missing. It is needed to consume external services`);
+                logger.warn(`${constants.EXT_MODULE} - ${constants.SAP_HTTP_CLIENT} dependency is missing. It is needed to consume external services`);
             }
             resolve();
         })
     } else {
-        logger.warn(`/${constants.PACKAGE} does not exist, please initialize a CAP Project`);
+        logger.warn(`${constants.EXT_MODULE} - /${constants.PACKAGE} does not exist, please initialize a CAP Project`);
     }
 
 }
@@ -66,13 +66,13 @@ let writeExtSrv = function (target) {
             fs.readFile(path.join(__dirname, '..', constants.SRV_FOLDER, constants.EXT_SERVICE_CDS), 'utf-8', (err, data) => {
                 fs.writeFile(path.join(target, constants.SRV_FOLDER, constants.EXT_SERVICE_CDS), data, (err) => {
                     if (err) console.error(err);
-                    logger.info(`/${constants.SRV_FOLDER}/${constants.EXT_SERVICE_CDS} created`)
+                    logger.info(`${constants.EXT_MODULE} - /${constants.SRV_FOLDER}/${constants.EXT_SERVICE_CDS} created`)
                     resolve();
                 })
             })
         })
     } else {
-        logger.info(`/${constants.SRV_FOLDER}/${constants.EXT_SERVICE_CDS} exists, skipping creation`)
+        logger.info(`${constants.EXT_MODULE} - /${constants.SRV_FOLDER}/${constants.EXT_SERVICE_CDS} exists, skipping creation`)
     }
 }
 
@@ -85,13 +85,13 @@ let writeExtSrvImpl = function (target) {
             fs.readFile(path.join(__dirname, '..', constants.SRV_FOLDER, constants.EXT_SERVICE_JS), 'utf-8', (err, data) => {
                 fs.writeFile(path.join(target, constants.SRV_FOLDER, constants.EXT_SERVICE_JS), data, (err) => {
                     if (err) console.error(err);
-                    logger.info(`/${constants.SRV_FOLDER}/${constants.EXT_SERVICE_JS} created`)
+                    logger.info(`${constants.EXT_MODULE} - /${constants.SRV_FOLDER}/${constants.EXT_SERVICE_JS} created`)
                     resolve();
                 })
             })
         })
     } else {
-        logger.info(`/${constants.SRV_FOLDER}/${constants.EXT_SERVICE_JS} exists, skipping creation`)
+        logger.info(`${constants.EXT_MODULE} - /${constants.SRV_FOLDER}/${constants.EXT_SERVICE_JS} exists, skipping creation`)
     }
 }