+++ utils/ui5.js
@@ -12,7 +12,7 @@ const constants = require('./constants');
 let createApp = async function (target) {
     return new Promise(async (resolve, reject) => {
         try {
-            logger.info(`Copying ${constants.APP_FOLDER} folder and its content...`);
+            logger.info(`${constants.UI5_MODULE} - Copying ${constants.APP_FOLDER} folder and its content...`);
 
             await fs.ensureDir(path.join(target,constants.APP_FOLDER));
 
@@ -21,10 +21,10 @@ let createApp = async function (target) {
                             path.join(target,constants.APP_FOLDER)
                         );
         
-            logger.info(`Folder ${constants.APP_FOLDER} and its content copied`);
+            logger.info(`${constants.UI5_MODULE} - Folder ${constants.APP_FOLDER} and its content copied`);
             resolve();
           } catch (err) {
-            logger.error('Error during folder copy:', err);
+            logger.error('${constants.UI5_MODULE} - Error during folder copy:', err);
             reject(err);
           }
     })